added heart rate averaging to the main loop
- this should probably be a function in the heart rate library that returns a new value if one is available when called and returns the prev value if not

+++ 00_PIC/team207_main_testing.X/heartrate.h
@@ -37,7 +37,7 @@ extern "C" {
 // - 2: Polling complete
 struct HrResults {
     uint8_t status;
-    unsigned int hr;
+    float hr; // float heart rate that is to be averaged over a number of samples
 };
    
 
@@ -48,7 +48,7 @@ struct HrResults pollHR(uint8_t mode);
 void triggerHR();
 
 // Placeholder function for beat detection
-unsigned int detectBeats(float threshold, int time);
+float detectBeats(float threshold, int time);
 
 int arrayLen;
 //static float dataArray[32];
